server/requirements.txt
# python3.8.5
Django==3.1.3
djangorestframework==3.12.1
django-dramatiq==0.9.1
dramatiq[redis, watch]==1.9.0
tushare==1.2.62
Pandas3==0.0.1
django-apscheduler==0.5.0
pypinyin==0.39.1
psycopg2==2.8.6
django-redis==4.12.1
daphne==3.0.1